backend/pyproject.toml
[project]
name = "rag-fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.9",
    "fastapi[standard]>=0.115",
    "ollama>=0.4.8",
    "pydantic-settings>=2.9.1",
    "sentence-transformers>=4.1.0",
]
